pyproject.toml
[tool.poetry]
name = "wbseo"
version = "0.1.0"
description = ""
authors = ["legkovoyg <147488090+legkovoyg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
openai = "^1.63.0"
requests = "^2.32.3"
redis = "^5.2.1"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
